Document AssistantForm refactor and relocate non-refactored sharing components
- Add/update refactor.md: explicit out-of-scope section for assistant sharing,
  correct file index and metrics (10 form components), remove sharing from the
  form Data flow mermaid to avoid implying they were refactored with the form.
- Move AssistantSharing.svelte and AssistantSharingModal.svelte to assistants/
  root (same level as AssistantForm.svelte).
- Update route imports and documentation paths that referenced components/.
