+++ frontend/svelte-app/src/lib/components/assistants/AssistantSharing.svelte
