+++ frontend/svelte-app/src/lib/components/assistants/AssistantSharingModal.svelte
